requirements.txt
albumentations==0.4.3
DISTS_pytorch==0.1
einops==0.8.0
numpy==1.24.4; python_version<'3.12'
numpy==1.26.4; python_version>='3.12' and python_version<'3.13'
numpy==2.1.*;  python_version>='3.13'
pandas==2.2.2
openpyxl==3.1.5
omegaconf==2.0.6
opencv_python_headless==4.9.0.80
Pillow==10.3.0
pytorch_lightning==2.2.4
pytorch_msssim==1.0.0
PyYAML==6.0.1
requests==2.31.0
timm==1.0.3
torch==2.2.0
torchac==0.9.3
torchvision==0.17.0
torch-fidelity==0.3.0
tqdm==4.66.4
transformers==4.38.2
typing_extensions==4.11.0
lpips==0.1.4
piq==0.8.0
ninja
tensorboard
open_clip_torch
faiss-cpu
zstandard
fastapi>=0.111
uvicorn[standard]>=0.30
python-multipart
requests